IRGen: always use runtime initialized VWT on PE/COFF
Use the lazy initialization path for PE/COFF which cannot have static
references to module external symbols without an indirection.  This
results in the inability to have constant data with references to module
external as we do in the VWT when chaining up to the parent type (e.g
NativeObject).