Generate and Improve Aligned Structs in TypeLayout
Rather than deferring to TypeInfo when generating type layouts, create
an aligned struct instead. Since the alignment computation in the
default codegen for aligned structs produces a lot more code than the
typeinfo route, also enhanced the codegen for fixed sized structs.

If the Aligned Group is fixed sized and alignment, rather than doing an
alignment computation at runtime, we compute the offset for each field
and emit a GEP.